script to split the dataset and apply GLMM
The script will split the dataset into 10 subset used during the k-fold cross-validation. It will export in the folder different documents: 
the GLMM model summary, the ROC curve for each of the 10folds, the confusion matrix for each of the 10folds. It will also export a .csv document with all the observations and predictions